Add JS version two
I managed to get the latitude and longitude on my laptop but still need
to use the secured API provided by FreeCodeCamp to get the users
location. I also created a function with errors which is the second
argument for getCurrentPosition method.  I am trying to avoid using var
and instead use cont and let.