Add JS version seven
I am using vanilla JS to toggle between Celsius and Fahrenheit and I
added global variable for accessData instead of local variable and
assign it inside the function. Also added a new global variable and set
it to false. Created an event listener to tempEl and called toggleTemp
in the callback function. Also I tried to add comments to my codes.