Add JS version eight
I changed the condition from main to description and have it in
uppercase and now I am trying to do the icons with Skycons library.
Still working on the icons and then I have few things to check and add
like the loading bar before the data are displayed on screen.