Add JS version nine
I succeeded in adding the Skycons library to my app and used a switch
statement to cover the conditions I might have. Still thinking about
how to add the partly_cloudy_day and the partly_cloudy_night. I think I
will need to link them to the sunrise and sunset in the json response.
Also I want to check what are the different values for the condition
but still trying to figure that out as well.