Add some styling
I am trying to have a design for the app but I am still working on it.
I am starting with the mobile 1st design then scale it to tablets and
desktops. I also commented out the loader for now until I figure out
how to just have it when the user click on allow to show the location
and then the loader starts and it is disabled when the data are
downloaded.